chore: fix all typing hint

+++ src/linear_function.py
@@ -1,4 +1,4 @@
-def linear_function(theta0: float, theta1: float, x: int) -> float:
+def linear_function(theta0: float, theta1: float, x: float) -> float:
     """
     linear_function(theta0: float, theta1: float, x: int) -> float
     perform a linear regression (t1 * x) + t0
